Move CxxLocation to its own files.
