+++ ct/CxxLocation.cpp
@@ -1,0 +1,92 @@
+//==============================================================================
+//
+//  CxxLocation.cpp
+//
+//  Copyright (C) 2013-2020  Greg Utas
+//
+//  This file is part of the Robust Services Core (RSC).
+//
+//  RSC is free software: you can redistribute it and/or modify it under the
+//  terms of the GNU General Public License as published by the Free Software
+//  Foundation, either version 3 of the License, or (at your option) any later
+//  version.
+//
+//  RSC is distributed in the hope that it will be useful, but WITHOUT ANY
+//  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+//  FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+//  details.
+//
+//  You should have received a copy of the GNU General Public License along
+//  with RSC.  If not, see <http://www.gnu.org/licenses/>.
+//
+#include "CxxLocation.h"
+#include <string>
+
+using std::string;
+
+//------------------------------------------------------------------------------
+
+namespace CodeTools
+{
+CxxLocation::CxxLocation() :
+   file_(nullptr),
+   pos_(NOT_IN_SOURCE),
+   erased_(false),
+   internal_(false)
+{
+}
+
+//------------------------------------------------------------------------------
+
+size_t CxxLocation::GetPos() const
+{
+   if(erased_) return string::npos;
+   return (pos_ != NOT_IN_SOURCE ? pos_ : string::npos);
+}
+
+//------------------------------------------------------------------------------
+
+void CxxLocation::SetLoc(CodeFile* file, size_t pos)
+{
+   file_ = file;
+   pos_ = pos;
+}
+
+//------------------------------------------------------------------------------
+
+void CxxLocation::UpdatePos(EditorAction action,
+   size_t begin, size_t count, size_t from)
+{
+   switch(action)
+   {
+   case Erased:
+      if(!erased_ && (pos_ >= begin))
+      {
+         if(pos_ < (begin + count))
+            erased_ = true;
+         else
+            pos_ -= count;
+      }
+      break;
+
+   case Inserted:
+      if(!erased_ && (pos_ >= begin)) pos_ += count;
+      break;
+
+   case Pasted:
+      if(erased_)
+      {
+         if((pos_ >= from) && (pos_ < (from + count)))
+         {
+            pos_ = pos_ + begin - from;
+            erased_ = false;
+         }
+      }
+      else
+      {
+         if(pos_ >= begin) pos_ += count;
+      }
+      break;
+   }
+}
+}
