+++ ct/CxxNamed.cpp
@@ -60,70 +60,6 @@ CxxScoped* ReferentError(const string& item, debug64_t offset)
 
 //==============================================================================
 
-CxxLocation::CxxLocation() :
-   file_(nullptr),
-   pos_(NOT_IN_SOURCE),
-   erased_(false),
-   internal_(false)
-{
-}
-
-//------------------------------------------------------------------------------
-
-size_t CxxLocation::GetPos() const
-{
-   if(erased_) return string::npos;
-   return (pos_ != NOT_IN_SOURCE ? pos_ : string::npos);
-}
-
-//------------------------------------------------------------------------------
-
-void CxxLocation::SetLoc(CodeFile* file, size_t pos)
-{
-   file_ = file;
-   pos_ = pos;
-}
-
-//------------------------------------------------------------------------------
-
-void CxxLocation::UpdatePos(EditorAction action,
-   size_t begin, size_t count, size_t from)
-{
-   switch(action)
-   {
-   case Erased:
-      if(!erased_ && (pos_ >= begin))
-      {
-         if(pos_ < (begin + count))
-            erased_ = true;
-         else
-            pos_ -= count;
-      }
-      break;
-
-   case Inserted:
-      if(!erased_ && (pos_ >= begin)) pos_ += count;
-      break;
-
-   case Pasted:
-      if(erased_)
-      {
-         if((pos_ >= from) && (pos_ < (from + count)))
-         {
-            pos_ = pos_ + begin - from;
-            erased_ = false;
-         }
-      }
-      else
-      {
-         if(pos_ >= begin) pos_ += count;
-      }
-      break;
-   }
-}
-
-//==============================================================================
-
 Asm::Asm(ExprPtr& code) : code_(std::move(code))
 {
    Debug::ft("Asm.ctor");
