+++ ct/CxxNamed.h
@@ -30,6 +30,7 @@
 #include "CodeTypes.h"
 #include "Cxx.h"
 #include "CxxFwd.h"
+#include "CxxLocation.h"
 #include "CxxString.h"
 #include "SysTypes.h"
 
@@ -37,86 +38,6 @@
 
 namespace CodeTools
 {
-//  Where a C++ item was declared or defined.
-//
-class CxxLocation
-{
-   friend class CxxNamed;
-   friend class CxxStatement;
-public:
-   //  Returns the file in which the item is located.  A template
-   //  instance belongs to the file that caused its instantiation.
-   //  An item added by the Editor belongs to the file to which it
-   //  was added.
-   //
-   CodeFile* GetFile() const { return file_; }
-
-   //  Returns the start of the item's position within its file, which
-   //  is an index into a string that contains the file's contents.
-   //  For a template instance, this is an offset into its internally
-   //  generated code.  For an item added by the Editor, string::npos
-   //  is returned.
-   //
-   size_t GetPos() const;
-
-   //  Returns true for an internally generated item, such as the code
-   //  for a template instance.
-   //
-   bool IsInternal() const { return internal_; }
-
-   //  Updates the item's location after code has been edited.  Has the
-   //  same interface as CxxToken::UpdatePos.
-   //
-   void UpdatePos(EditorAction action, size_t begin, size_t count, size_t from);
-
-   //  A position that indicates that the item was not found in the
-   //  original source code.  This is used when the Editor creates
-   //  an item, for example.
-   //
-   static const size_t NOT_IN_SOURCE = 0x3fffffff;
-private:
-   //  Initializes fields to default values.  Private because this class
-   //  only appears as a private member of its friend classes.
-   //
-   CxxLocation();
-
-   //  Copy constructor.
-   //
-   CxxLocation(const CxxLocation& that) = default;
-
-   //  Copy operator.
-   //
-   CxxLocation& operator=(const CxxLocation& that) = default;
-
-   //  Records the item's location in source code.
-   //
-   void SetLoc(CodeFile* file, size_t pos);
-
-   //  Marks the item as internally generated.
-   //
-   void SetInternal() { internal_ = true; }
-
-   //  The file in which the item appeared.
-   //
-   CodeFile* file_;
-
-   //  The item's location in FILE.  The file has a string member which
-   //  contains the code, and this is an index into that string.
-   //
-   size_t pos_ : 30;
-
-   //  Set if the item has been erased during editing.
-   //
-   bool erased_ : 1;
-
-   //  Set if the item appeared in internally generated code, which currently
-   //  means in a template instance.
-   //
-   bool internal_ : 1;
-};
-
-//------------------------------------------------------------------------------
-//
 //  The base class for C++ entities that define a name.  Each of these knows
 //  the file in which it was found.
 //
