+++ ct/ct.vcxproj
@@ -143,6 +143,7 @@
     <ClCompile Include="CxxArea.cpp" />
     <ClCompile Include="CxxDirective.cpp" />
     <ClCompile Include="CxxExecute.cpp" />
+    <ClCompile Include="CxxLocation.cpp" />
     <ClCompile Include="CxxNamed.cpp" />
     <ClCompile Include="CxxRoot.cpp" />
     <ClCompile Include="CxxScope.cpp" />
@@ -181,6 +182,7 @@
     <ClInclude Include="CxxDirective.h" />
     <ClInclude Include="CxxExecute.h" />
     <ClInclude Include="CxxFwd.h" />
+    <ClInclude Include="CxxLocation.h" />
     <ClInclude Include="CxxNamed.h" />
     <ClInclude Include="CxxRoot.h" />
     <ClInclude Include="CxxScope.h" />
