+++ ct/ct.vcxproj.filters
@@ -114,6 +114,9 @@
     <ClCompile Include="CtIncrement.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="CxxLocation.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="SourceCode.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -225,6 +228,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="CtIncrement.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CxxLocation.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="SourceCode.h">
